Reindex process queues project image operations.
