+++ common-lib/src/main/scala/com/gu/mediaservice/model/ThrallMessage.scala
@@ -90,6 +90,7 @@ object ExternalThrallMessage{
   implicit val upsertFromProjectionMessage: OFormat[UpsertFromProjectionMessage] = Json.format[UpsertFromProjectionMessage]
 
   implicit val createInstanceMessage: OFormat[CreateInstanceMessage] = Json.format[CreateInstanceMessage]
+  implicit val reindexImageMessage: OFormat[ReindexImageMessage] = Json.format[ReindexImageMessage]
 
   implicit val writes: OWrites[ExternalThrallMessage] = Json.writes[ExternalThrallMessage]
   implicit val reads: Reads[ExternalThrallMessage] = Json.reads[ExternalThrallMessage]
@@ -171,3 +172,5 @@ case class CompleteMigrationMessage(lastModified: DateTime, instance: Instance) 
 }
 
 case class CreateInstanceMessage(id: String, lastModified: DateTime, instance: Instance) extends ExternalThrallMessage
+
+case class ReindexImageMessage(id: String, lastModified: DateTime, instance: Instance) extends ExternalThrallMessage
